FeAmGen.jl Documentation

FeAmGen.jl Documentation

FeAmGen Functions in Message.jl

FeAmGen Functions in Extra.jl

FeAmGen Dict and Struct in Universe.jl

FeAmGen functions in SimpleDigest.jl

Index

FeAmGen Functions in Message.jl

FeAmGen.show_welcome_message — Function

show_welcome_message()

This function gives information for users. Return nothing.

FeAmGen.green_message — Function

green_message( statement::String, message::String )::Nothing

Print the statement in normal color and message in green. Return nothing.

FeAmGen Functions in Extra.jl

FeAmGen.degree — Function

degree( expr::Basic, x::Basic; max_n::Int64=100 )::Int64

A little extension for SymEngine. This function degree calculate the degree of polynomial expr of x. For now we do not have toolkit to constraint expr as polynomial, so we only choose a maximum degree in the loop.

FeAmGen.convert_to_array — Function

convert_to_array( ::Val{:Symbol}, mom::Basic )::Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}}}

This specific convertion is applied only on single symbol momentum, e.g. k1.

Examples

julia> using SymEngine, FeAmGen

julia> @vars k1
(k1,)

julia> FeAmGen.convert_to_array(k1)
1-element Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}},1}:
 (num = 1, ki = k1)

convert_to_array( ::Val{:Mul}, mom::Basic )::Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}}}

This specific convertion is applied only on non-unit coefficient with single symbol momentum, e.g. 2*k1.

Examples

julia> using SymEngine, FeAmGen

julia> @vars k1
(k1,)

julia> FeAmGen.convert_to_array(2*k1)
1-element Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}},1}:
 (num = 2, ki = k1)

convert_to_array( ::Val{:Add}, mom::Basic )::Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}}}

This specific convertion is applied only on the summation of non-unit coefficient with single symbol momentum, e.g. 2*k1+3*k2+k3.

Examples

julia> using SymEngine, FeAmGen

julia> @vars k1, k2, k3
(k1, k2, k3)

julia> FeAmGen.convert_to_array(2*k1+3*k2+k3)
3-element Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}},1}:
 (num = 1, ki = k3)
 (num = 2, ki = k1)
 (num = 3, ki = k2)

convert_to_array( mom::Basic )::Array{NamedTuple{(:num, :ki),Tuple{Basic,Basic}}}

This is the interface to different specific convertion functions according to the class of mom. Now we assume there only there different classes of momentum mom. For example,

k1

2*k1

2*k1+3*k2+k3

FeAmGen.make_SP — Function

make_SP( mom1::Basic, mom2::Basic)::Basic

This function linearly expands the scalar product of mom1 and mom2.

Examples

julia> using SymEngine, FeAmGen

julia> @vars k1, k2, k3
(k1, k2, k3)

julia> FeAmGen.make_SP( 2*k1+3*k2+k3, k1+k2 )
2*SP(k1, k1) + 2*SP(k1, k2) + 3*SP(k2, k1) + 3*SP(k2, k2) + SP(k3, k1) + SP(k3, k2)

FeAmGen.gen_mma_str — Function

gen_mma_str( expr::Basic )::String

This is a generic interface for different classes of the expr. And it will generate the mathematica code for the expression expr.

FeAmGen Dict and Struct in Universe.jl

FeAmGen.color_dict — Constant

const Dict color_dict

Dict( 1=>:singlet, 3=>:triplet, -3=>:triplet, 8=>:octet )

FeAmGen.spin_dict — Constant

const Dict spin_dict

Dict( 1=>:scalar, -1=>:ghost, 2=>:fermion, 3=>:vector )

FeAmGen.charge_convert — Function

charge_convert( model_charge::Float64 )::Rational{Int64}

Convert the float number of charge in the model file into a rational number, especially the quark charge.

FeAmGen.Particle — Type

struct Particle

  kf::Int64
  name::String
  antiname::String
  spin::Symbol
  color::Symbol
  charge::Rational{Int64}
  mass::Basic
  width::Basic

FeAmGen.Interaction — Type

struct Interaction

  name::String
  # UFO convention is outgoing particle, but in our convention is incoming
  link_list::Vector{Particle}
  color_row_list::Vector{Basic} # 1xN
  lorentz_col_list::Vector{Basic} # Mx1
  couplings_matrix::Array{Basic,2} # NxM
  QCD_order::Int64
  QED_order::Int64

FeAmGen.Model — Type

struct Model

  name::String
  unitary_gauge::Bool
  particle_list::Vector{Particle}
  particle_name_dict::Dict{String,Particle}
  particle_kf_dict::Dict{Int64,Particle}
  interaction_list::Vector{Interaction}
  sorted_kf_list_dict::Dict{Vector{Int64},Interaction}
  parameter_dict::Dict{Basic,Basic}

FeAmGen.is_anticommuting — Function

is_anticommuting( part::Particle )::Bool

Function "is_anticommuting" is used for the sign in the calculation of fermion/ghost loop. The problem is according to Paulo Nogueira, we can have our own definition for the sign of majorana loop.

FeAmGen.is_neutral — Function

is_neutral( part::Particle )::Bool

FeAmGen.is_not_majorana — Function

is_not_majorana( part::Particle )::Bool

FeAmGen.is_massless — Function

is_massless( part::Particle )::Bool

FeAmGen.is_massive — Function

is_massive( part::Particle )::Bool

FeAmGen.is_gluon — Function

is_gluon( part::Particle )::Bool

FeAmGen.is_photon — Function

is_photon( part::Particle )::Bool

FeAmGen.is_massless_quark — Function

is_massless_quark( part::Particle )::Bool

FeAmGen.is_top_quark — Function

is_top_quark( part::Particle )::Bool

FeAmGen.is_ghost — Function

is_ghost( part::Particle )::Bool

FeAmGen.is_quark — Function

is_quark( part::Particle )::Bool

FeAmGen.is_colorful — Function

is_colorful( part::Particle )::Bool

FeAmGen.is_not_colorful — Function

is_not_colorful( part::Particle )::Bool

FeAmGen.is_massive_fermion — Function

is_massive_fermion( part::Particle )::Bool

FeAmGen.to_string — Function

to_string( inter::Interaction )::String

Convert Interaction to string for writing out.

FeAmGen.find_AC_pair_pos — Function

find_AC_pair_pos( part_list::Vector{Particle} )::Union{Tuple{Nothing,Nothing},Tuple{Int64,Int64}}

This function is used to find the pair of anti-commuting particles/fields. This information will be inserted into QGRAF model, so that QGRAF can generate correct sign for anti-commuting particle loop. For example, we should have extra minus sign for fermion loop and ghost loop.

FeAmGen.generate_ordered_link_list — Function

generate_ordered_link_list( part_list::Vector{Particle} )::Vector{Particle}

This function is used to generate order particle list according to the rule of QGRAF.

FeAmGen.to_qgraf_name — Function

to_qgraf_name( name_str::String )::String

Change the particle name into the one that QGRAF can accept.

FeAmGen.generate_QGRAF_model — Function

generate_QGRAF_model( model::Model )::Nothing

Write out the model file that can be used by QGRAF.

FeAmGen.logging_model — Function

logging_model( model::Model )::Nothing

Write out model file that contains more detail.

FeAmGen functions in SimpleDigest.jl

FeAmGen.electron_num — Function

electron_num( part::Particle )::Int64

Count the electron number (lepton number in 1st generation ).

FeAmGen.mu_num — Function

mu_num( part::Particle )::Int64

Count the muon number (lepton number in 2nd generation ).

FeAmGen.tau_num — Function

tau_num( part::Particle )::Int64

Count the tauon number (lepton number in 3rd generation ).

FeAmGen.quark1st_num — Function

quark1st_num( part::Particle )::Int64

Count the quark flavor number in 1st generation. This is used for the case that has no flavor violation.

FeAmGen.quark2nd_num — Function

quark2nd_num( part::Particle )::Int64

Count the quark flavor number in 2nd generation. This is used for the case that has no flavor violation.

FeAmGen.quark3rd_num — Function

quark3rd_num( part::Particle )::Int64

Count the quark flavor number in 3rd generation. This is used for the case that has no flavor violation.

Index

FeAmGen.color_dict

FeAmGen.spin_dict

FeAmGen.Interaction

FeAmGen.Model

FeAmGen.Particle

FeAmGen.charge_convert

FeAmGen.convert_to_array

FeAmGen.degree

FeAmGen.electron_num

FeAmGen.find_AC_pair_pos

FeAmGen.gen_mma_str

FeAmGen.generate_QGRAF_model

FeAmGen.generate_ordered_link_list

FeAmGen.green_message

FeAmGen.is_anticommuting

FeAmGen.is_colorful

FeAmGen.is_ghost

FeAmGen.is_gluon

FeAmGen.is_massive

FeAmGen.is_massive_fermion

FeAmGen.is_massless

FeAmGen.is_massless_quark

FeAmGen.is_neutral

FeAmGen.is_not_colorful

FeAmGen.is_not_majorana

FeAmGen.is_photon

FeAmGen.is_quark

FeAmGen.is_top_quark

FeAmGen.logging_model

FeAmGen.make_SP

FeAmGen.mu_num

FeAmGen.quark1st_num

FeAmGen.quark2nd_num

FeAmGen.quark3rd_num

FeAmGen.show_welcome_message

FeAmGen.tau_num

FeAmGen.to_qgraf_name

FeAmGen.to_string